+++ services/ui-src/src/measures/2023/shared/util/validationsMock.tsx
@@ -8,7 +8,7 @@ import * as validateBothDatesInRange from "shared/globalValidations/validateBoth
 import * as validateDualPopInformation from "measures/2023/shared/globalValidations/validateDualPopInformation";
 import * as validateEqualCategoryDenominators from "measures/2023/shared/globalValidations/validateEqualCategoryDenominators";
 import * as validateEqualQualifierDenominators from "shared/globalValidations/validateEqualQualifierDenominators";
-import * as validateFfsRadioButtonCompletion from "measures/2023/shared/globalValidations/validateFfsRadioButtonCompletion";
+import * as validateFfsRadioButtonCompletion from "shared/globalValidations/validateFfsRadioButtonCompletion";
 import * as validateRateNotZero from "measures/2023/shared/globalValidations/validateRateNotZero";
 import * as validateRateZero from "measures/2023/shared/globalValidations/validateRateZero";
 import * as validateNumeratorsLessThanDenominators from "shared/globalValidations/validateNumeratorsLessThanDenominators";
