+++ services/ui-src/src/shared/globalValidations/validateFfsRadioButtonCompletion/index.test.ts
@@ -1,4 +1,4 @@
-import { testFormData } from "../testHelpers/_testFormData";
+import { testFormData } from "measures/2024/shared/globalValidations/testHelpers/_testFormData";
 import * as DC from "dataConstants";
 import { validateFfsRadioButtonCompletion } from ".";
